Updated ORM

+++ packages/orm/angel_orm_postgres/test/migrations/person_order.sql
@@ -2,7 +2,7 @@ CREATE TEMPORARY TABLE "person_orders" (
   "id" serial PRIMARY KEY,
   "person_id" int not null,
   "name" varchar(255),
-  "price" int,
+  "price" float,
   "deleted" bool not null default false,
   "created_at" timestamp,
   "updated_at" timestamp
